Rename vuetify types file, add VTooltip
