+++ frontend/composables/recipes/use-recipe-tools.ts
@@ -1,7 +1,7 @@
 import { reactive, ref, useAsync } from "@nuxtjs/composition-api";
 import { useAsyncKey } from "../use-utils";
 import { useUserApi } from "~/composables/api";
-import { VForm } from "~/types/v-form";
+import { VForm } from "~/types/vuetify";
 
 export const useTools = function (eager = true) {
   const workingToolData = reactive({
