+++ frontend/pages/admin/manage/groups/_id.vue
@@ -31,7 +31,7 @@ import { useGroups } from "~/composables/use-groups";
 import { alert } from "~/composables/use-toast";
 import { useUserForm } from "~/composables/use-users";
 import { validators } from "~/composables/use-validators";
-import { VForm } from "~/types/v-form";
+import { VForm } from "~/types/vuetify";
 
 export default defineComponent({
   components: {
