+++ frontend/pages/recipe/create.vue
@@ -322,7 +322,7 @@ import RecipeCategoryTagSelector from "~/components/Domain/Recipe/RecipeCategory
 import { validators } from "~/composables/use-validators";
 import { Recipe } from "~/types/api-types/recipe";
 import { alert } from "~/composables/use-toast";
-import { VForm} from "~/types/v-form";
+import { VForm} from "~/types/vuetify";
 
 export default defineComponent({
   components: { VJsoneditor, RecipeCategoryTagSelector },
