+++ frontend/pages/user/profile/api-tokens.vue
@@ -66,7 +66,7 @@
 <script lang="ts">
 import { computed, defineComponent, useContext, ref } from "@nuxtjs/composition-api";
 import { useUserApi } from "~/composables/api";
-import { VForm } from "~/types/v-form";
+import { VForm } from "~/types/vuetify";
 
 export default defineComponent({
   setup() {
