+++ frontend/types/vuetify.ts
@@ -1,3 +1,9 @@
+// TODO Remove this file when upgrading to Vuetify 3.0
+
+export type VTooltip = Vue & {
+  deactivate(): void;
+}
+
 export type VForm = Vue & {
   validate: () => boolean;
   resetValidation: () => boolean;
